Update README.md
In this recitation assignment, you are given the definition of a class, called Value, which represents values of operands in the SPL language to implement its interpreter in Programming Assignment 3. These include values for the two defined types in the language: Numeric (double), and String (string). The objective of defining the Value class is to facilitate constructing an interpreter for the SPL language which evaluates expressions and executes statements using C++.

In RA 8, you are required to implement some of the member functions of the Value class that apply the numeric or string operators in order to enable testing the class separately as a unit before using it in the construction of the interpreter in PA3. You are required to implement the following member functions:

operator*(), operator==(), operator^(), Catenate(), SLthan().

The semantic rules governing the evaluation of expressions in the SPL language are summarized below: